Navigation (#28)
* Upgrade all dependencies

* Add basis setup for Navigation

* Add SafeArgs setup

* Update README.md

* Rename Arch dependencies

* Migrated from fabrics to fb crashlytics

+++ hooks/post_gen_project.sh
@@ -3,6 +3,7 @@
 APP_NAME="{{ cookiecutter.app_name }}"
 USE_ANALYTICS="{{ cookiecutter.firebase_analytics }}"
 USE_FB_MESSAGING="{{ cookiecutter.firebase_messaging }}"
+USE_FB_CRASHLYTICS="{{ cookiecutter.firebase_crashlytics }}"
 STRING_TOOL="{{ cookiecutter.string_tool }}"
 PACKAGE_NAME="{{ cookiecutter.package_name }}"
 REPO_NAME="{{ cookiecutter.repo_name }}"
@@ -17,7 +18,7 @@ cd ..
 
 echo "*) GIT setup completed"
 
-if [[ $USE_ANALYTICS == "yes" || $USE_FB_MESSAGING == "yes" ]]
+if [[ $USE_ANALYTICS == "yes" || $USE_FB_MESSAGING == "yes" || $USE_FB_CRASHLYTICS == "yes" ]]
 then
     if [[ $USE_ANALYTICS == "yes"  ]]
     then
@@ -27,18 +28,21 @@ then
     then
         echo "*) Firebase Messaging setup successful"
     fi
+    if [[ $USE_FB_CRASHLYTICS == "yes"  ]]
+    then
+        echo "*) Firebase Crashlytics setup successful"
+    fi
 
     echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
     echo "Please do not forget to modify google-services.json file!!!!"
     echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
 fi
 
-
 if [[ $USE_FB_MESSAGING != "yes" ]]
 then
     echo "*) Unused FCM folders and files removed"
     cd {{ cookiecutter.repo_name }}/app
-    if [[ $USE_ANALYTICS != "yes" ]]
+    if [[ $USE_ANALYTICS != "yes" && $USE_FB_CRASHLYTICS != "yes" ]]
     then
         rm google-services.json
     fi
@@ -51,6 +55,7 @@ fi
 
 if [[ $STRING_TOOL != "texterify" ]]
 then
+    cd $ROOT_DIRECTORY
     cd app
     rm texterify.json
     cd $ROOT_DIRECTORY #back to project root
